+++ index.html
@@ -226,9 +226,11 @@
       </div>
     </section>
   </template>
-
-  <script src="js/function.js"></script>
-  <script src="/js/main.js"></script>
-
+    <script type="module" src="main.js"></script>
+    <script type="module" src="comment.js"></script>
+    <script type="module" src="comments.js"></script>
+    <script type="module" src="function.js"></script>
+    <script type="module" src="photo.js"></script>
+    <script type="module" src="utils.js"></script>
 </body>
 </html>
